move retriever test to proper pkg
